+++ src/mordordefense/Saruman.java
@@ -11,14 +11,14 @@ public class Saruman {
 	/**
 	 * Szarumán varázsereje.
 	 */
-	private int manna;
+	private float manna;
 
 	/**
 	 * Szarumán konstruktora
 	 * 
 	 * @param mana
 	 */
-	public Saruman(int mana) {
+	public Saruman(float mana) {
 		Logging.log(2, ">> Saruman konstruktor hívás, paraméter: " + mana);
 		manna = mana;
 		Logging.log(4, "<< Saruman konstruktor");
@@ -30,7 +30,7 @@ public class Saruman {
 	 * @param m
 	 *            A hozzáadódó varázserő
 	 */
-	public void addManna(int m) {
+	public void addManna(float m) {
 		Logging.log(1, ">> Saruman.addManna() hívás, paraméter: " + m);
 		manna += m;
 		Logging.log(4, "<< Saruman.addManna");
@@ -41,7 +41,7 @@ public class Saruman {
 	 * 
 	 * @return Szarumán varázsereje.
 	 */
-	public int getManna() {
+	public float getManna() {
 		Logging.log(3, ">> Saruman.getManna() hívás.");
 		Logging.log(3, "<< " + manna);
 		return manna;
@@ -53,7 +53,7 @@ public class Saruman {
 	 * @param m
 	 *            Mennyivel csökkenjen a varázserő.
 	 */
-	public void rmManna(int m) {
+	public void rmManna(float m) {
 		Logging.log(1, ">> Saruman.rmManna() hívás, paraméter: " + m);
 		manna -= m;
 		Logging.log(3, "\t Szarumán új varázserej: " + manna);
